# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contracts API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Base server"
        }
    ],
    "paths": {
        "/api/audit-logs": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "აუდიტის ჟურნალი - სიის ნახვა",
                "operationId": "8a24c03b06fe3071c4035f4176892716",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "თარიღი - საწყისი (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "თარიღი - საბოლოო",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "table",
                        "in": "query",
                        "description": "ცხრილის სახელი",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "description": "მოქმედება (create/update/delete/upload/unlink)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/audit-logs/export": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "აუდიტ-რეპორტი ექსპორტი CSV/JSON",
                "operationId": "299cd68c6a859cd284360cc0b9444c26",
                "parameters": [
                    {
                        "name": "format",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "csv",
                                "json"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "table",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/contractors": {
            "get": {
                "tags": [
                    "Contractors"
                ],
                "summary": "კონტრაგენტების სიის ნახვა/ძიება (სწრაფი შევსებისთვის)",
                "operationId": "ba6e97581a2f26840c9d7a4617df5da5",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "ძიება სახელით/ტელეფონით/ელფოსტით",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "რამდენი ჩანაწერი დავაბრუნოთ (default: 20, max: 100)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contractors"
                ],
                "summary": "კონტრაგენტის შექმნა",
                "operationId": "8e5ed6d296b0b24548697fc44e094841",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "დასახელება",
                                        "type": "string",
                                        "example": "შპს ახალი პარტნიორი"
                                    },
                                    "phone": {
                                        "description": "ტელეფონი (E.164)",
                                        "type": "string",
                                        "example": "+995595000003",
                                        "nullable": true
                                    },
                                    "email": {
                                        "description": "ელფოსტა",
                                        "type": "string",
                                        "example": "partner@example.com",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contractors/{id}": {
            "put": {
                "tags": [
                    "Contractors"
                ],
                "summary": "კონტრაგენტის რედაქტირება",
                "operationId": "5a6352a805d4178ca5c79e7c0dc5ca7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "დასახელება",
                                        "type": "string",
                                        "example": "შპს პარტნიორი განახლებული"
                                    },
                                    "phone": {
                                        "description": "ტელეფონი (E.164)",
                                        "type": "string",
                                        "example": "+995595000004",
                                        "nullable": true
                                    },
                                    "email": {
                                        "description": "ელფოსტა",
                                        "type": "string",
                                        "example": "updated@example.com",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contractors"
                ],
                "summary": "კონტრაგენტის წაშლა",
                "operationId": "240ad80e5ca75268bf361931e20e8b13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}